@@ -17,47 +17,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.0.1",
-    "@angular/common": "9.0.1",
-    "@angular/compiler": "9.0.1",
-    "@angular/core": "9.0.1",
-    "@angular/forms": "9.0.1",
-    "@angular/localize": "9.0.1",
-    "@angular/platform-browser": "9.0.1",
-    "@angular/platform-browser-dynamic": "9.0.1",
-    "@angular/router": "9.0.1",
-    "angular-in-memory-web-api": "0.9.0",
-    "auth0-js": "9.13.2",
+    "@angular/animations": "10.2.2",
+    "@angular/common": "10.2.2",
+    "@angular/compiler": "10.2.2",
+    "@angular/core": "10.2.2",
+    "@angular/forms": "10.2.2",
+    "@angular/localize": "10.2.2",
+    "@angular/platform-browser": "10.2.2",
+    "@angular/platform-browser-dynamic": "10.2.2",
+    "@angular/router": "10.2.2",
+    "angular-in-memory-web-api": "0.11.0",
+    "@auth0/auth0-angular": "1.2.0",
     "body-parser": "1.19.0",
     "bootstrap": "4.4.1",
     "core-js": "3.6.4",
     "debug": "4.1.1",
     "express": "4.17.1",
+    "express-jwt": "6.0.0",
     "fast-sort": "2.0.1",
     "jquery": "3.5.0",
+    "jwks-rsa": "1.11.0",
     "moment": "2.24.0",
     "morgan": "1.9.1",
-    "@ng-bootstrap/ng-bootstrap": "6.0.0-rc.0",
+    "@ng-bootstrap/ng-bootstrap": "8.0.0",
     "path": "0.12.7",
-    "pg": "7.17.1",
+    "pg": "8.4.2",
     "popper.js": "1.16.0",
-    "rxjs": "6.5.4",
-    "sequelize": "5.21.3",
+    "rxjs": "6.6.3",
+    "sequelize": "6.3.5",
     "socket.io": "2.3.0",
     "socket.io-client": "2.3.0",
-    "typescript": "3.7.5",
+    "typescript": "4.0.5",
     "underscore": "1.9.2",
-    "zone.js": "0.10.2"
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.901.12",
-    "@angular/cli": "9.0.2",
-    "@angular/compiler-cli": "9.0.1",
-    "@angular/language-service": "9.0.1",
+    "@angular-devkit/build-angular": "0.1002.0",
+    "@angular/cli": "10.2.0",
+    "@angular/compiler-cli": "10.2.2",
+    "@angular/language-service": "10.2.2",
     "@types/jasmine": "3.5.5",
     "@types/jasminewd2": "2.0.8",
     "@types/node": "13.1.7",
-    "codelyzer": "5.2.1",
+    "codelyzer": "6.0.1",
     "enhanced-resolve": "4.1.1",
     "jasmine-core": "3.5.0",
     "jasmine-spec-reporter": "4.2.1",
@@ -66,14 +68,13 @@
     "karma-coverage-istanbul-reporter": "2.1.1",
     "karma-jasmine": "3.1.0",
     "karma-jasmine-html-reporter": "1.5.1",
-    "ordinal": "1.0.3",
     "protractor": "7.0.0",
     "ts-node": "8.6.2",
     "tslib": "1.10.0",
     "tslint": "5.20.1"
   },
   "engines": {
-    "node": "12.14.1",
-    "npm": "6.13.4"
+    "node": "14.15.0",
+    "npm": "6.14.8"
   }
 }
