@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {ActiveContext} from '../categories.context';
-import {AuthService} from '../../services/auth/auth.service';
+import {MyAuthService} from '../../services/auth/my-auth.service';
 
 @Component({
   selector: 'osc-winner-main',
@@ -11,8 +11,8 @@ export class WinnerMainComponent implements OnInit {
   public activeContext = ActiveContext.Winner;
   public person;
 
-  constructor(private auth: AuthService) {
-    this.auth.getPerson().subscribe(person => this.person = person);
+  constructor(private auth: MyAuthService) {
+    this.auth.me$.subscribe(person => this.person = person);
   }
 
   ngOnInit() {
